Added page restrictions for non authenticated users

+++ studymate/base/views.py
@@ -1,7 +1,9 @@
 from django.contrib import messages
 from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
 from django.contrib.auth.models import User
 from django.db.models import Q
+from django.http import HttpResponse
 from django.shortcuts import render, redirect
 
 from base.forms import RoomForm
@@ -9,6 +11,9 @@ from base.models import Room, Topic
 
 
 def login_page(request):
+    if request.user.is_authenticated:
+        return redirect("home")
+
     if request.method == "POST":
         username = request.POST.get("username")
         password = request.POST.get("password")
@@ -18,9 +23,10 @@ def login_page(request):
         except User.DoesNotExist:
             messages.error(request, "Invalid credentials")
 
-        user = authenticate(username=username, password=password)
-        login(request, user)
-        return redirect("home")
+        user = authenticate(request, username=username, password=password)
+        if user:
+            login(request, user)
+            return redirect("home")
 
     return render(request, "base/login_register.html")
 
@@ -43,6 +49,7 @@ def home(request):
     return render(request, "base/home.html", {"rooms": rooms, "topics": topics, "room_count": room_count})
 
 
+@login_required(login_url="login")
 def create_room(request):
     form = RoomForm()
     if request.method == "POST":
@@ -59,9 +66,14 @@ def get_room(request, pk):
     return render(request, "base/room.html", context)
 
 
+@login_required(login_url="login")
 def update_room(request, pk):
     required_room = Room.objects.get(id=pk)
     form = RoomForm(instance=required_room)
+
+    if request.user != required_room.host and not (request.user.is_staff or request.user.is_super_user):
+        return HttpResponse("Forbidden")
+
     if request.method == "POST":
         form = RoomForm(request.POST, instance=required_room)
         if form.is_valid():
@@ -70,8 +82,13 @@ def update_room(request, pk):
     return render(request, "base/create_room.html", {"form": form})
 
 
+@login_required(login_url="login")
 def delete_room(request, pk):
     required_room = Room.objects.get(id=pk)
+
+    if request.user != required_room.host and not (request.user.is_staff or request.user.is_super_user):
+        return HttpResponse("Forbidden")
+
     if request.method == "POST":
         required_room.delete()
         return redirect("home")
